requirements.txt
python-dotenv
langchain
langgraph
langchain-groq
langchain-chroma
langchain-community
langchain-ollama
uuid
pandas
ddgs